requirements.txt
google-cloud-language>=2.0.0
flask>=1.1.2
gunicorn